Time: 5 ms (22.39%), Space: 6 MB (6.97%) - LeetHub

+++ 1802-maximum-value-at-a-given-index-in-a-bounded-array/1802-maximum-value-at-a-given-index-in-a-bounded-array.cpp
@@ -19,13 +19,12 @@ public:
         }else{
             sum+=((x*(n-j-1))-su(n-j-2));
         }
-        //cout<<k<<" "<<sum<<" "<<j<<endl;
         return sum<=ma;
     }
     int maxValue(int n, int index, int maxSum) {
-        int beg=1,end=maxSum,ans=0;
+        ll beg=1,end=maxSum,ans=0;
         while(beg<=end){
-            int mid=(beg+end)/2;
+            ll mid=(beg+end)/2;
             if(check(mid,n,maxSum,index)){
                 beg=mid+1;
                 ans=mid;
